+++ sdk/dotnet/GetPrivateLinkEndpointService.cs
@@ -57,10 +57,15 @@ namespace Pulumi.Mongodbatlas
     public sealed class GetPrivateLinkEndpointServiceResult
     {
         /// <summary>
-        /// Status of the interface endpoint.
+        /// Status of the interface endpoint for AWS.
         /// Returns one of the following values:
         /// </summary>
-        public readonly string ConnectionStatus;
+        public readonly string AwsConnectionStatus;
+        /// <summary>
+        /// Status of the interface endpoint for AZURE.
+        /// Returns one of the following values:
+        /// </summary>
+        public readonly string AzureStatus;
         /// <summary>
         /// Indicates if Atlas received a request to remove the interface endpoint from the private endpoint connection.
         /// </summary>
@@ -96,7 +101,9 @@ namespace Pulumi.Mongodbatlas
 
         [OutputConstructor]
         private GetPrivateLinkEndpointServiceResult(
-            string connectionStatus,
+            string awsConnectionStatus,
+
+            string azureStatus,
 
             bool deleteRequested,
 
@@ -120,7 +127,8 @@ namespace Pulumi.Mongodbatlas
 
             string providerName)
         {
-            ConnectionStatus = connectionStatus;
+            AwsConnectionStatus = awsConnectionStatus;
+            AzureStatus = azureStatus;
             DeleteRequested = deleteRequested;
             EndpointServiceId = endpointServiceId;
             ErrorMessage = errorMessage;