+++ sdk/go/mongodbatlas/privateLinkEndpointService.go
@@ -80,9 +80,12 @@ import (
 type PrivateLinkEndpointService struct {
 	pulumi.CustomResourceState
 
-	// Status of the interface endpoint.
+	// Status of the interface endpoint for AWS.
 	// Returns one of the following values:
-	ConnectionStatus pulumi.StringOutput `pulumi:"connectionStatus"`
+	AwsConnectionStatus pulumi.StringOutput `pulumi:"awsConnectionStatus"`
+	// Status of the interface endpoint for AZURE.
+	// Returns one of the following values:
+	AzureStatus pulumi.StringOutput `pulumi:"azureStatus"`
 	// Indicates if Atlas received a request to remove the interface endpoint from the private endpoint connection.
 	DeleteRequested pulumi.BoolOutput `pulumi:"deleteRequested"`
 	// Unique identifier of the interface endpoint you created in your VPC with the `AWS` or `AZURE` resource.
@@ -146,9 +149,12 @@ func GetPrivateLinkEndpointService(ctx *pulumi.Context,
 
 // Input properties used for looking up and filtering PrivateLinkEndpointService resources.
 type privateLinkEndpointServiceState struct {
-	// Status of the interface endpoint.
+	// Status of the interface endpoint for AWS.
 	// Returns one of the following values:
-	ConnectionStatus *string `pulumi:"connectionStatus"`
+	AwsConnectionStatus *string `pulumi:"awsConnectionStatus"`
+	// Status of the interface endpoint for AZURE.
+	// Returns one of the following values:
+	AzureStatus *string `pulumi:"azureStatus"`
 	// Indicates if Atlas received a request to remove the interface endpoint from the private endpoint connection.
 	DeleteRequested *bool `pulumi:"deleteRequested"`
 	// Unique identifier of the interface endpoint you created in your VPC with the `AWS` or `AZURE` resource.
@@ -172,9 +178,12 @@ type privateLinkEndpointServiceState struct {
 }
 
 type PrivateLinkEndpointServiceState struct {
-	// Status of the interface endpoint.
+	// Status of the interface endpoint for AWS.
 	// Returns one of the following values:
-	ConnectionStatus pulumi.StringPtrInput
+	AwsConnectionStatus pulumi.StringPtrInput
+	// Status of the interface endpoint for AZURE.
+	// Returns one of the following values:
+	AzureStatus pulumi.StringPtrInput
 	// Indicates if Atlas received a request to remove the interface endpoint from the private endpoint connection.
 	DeleteRequested pulumi.BoolPtrInput
 	// Unique identifier of the interface endpoint you created in your VPC with the `AWS` or `AZURE` resource.