+++ sdk/python/pulumi_mongodbatlas/get_private_endpoint.py
@@ -88,8 +88,8 @@ class GetPrivateEndpointResult:
     @pulumi.getter
     def status(self) -> str:
         """
-        Status of the AWS PrivateLink connection.
-        Returns one of the following values:
+        Status of the AWS PrivateLink connection or Status of the Azure Private Link Service. Atlas returns one of the following values:
+        AWS:
         """
         return pulumi.get(self, "status")